Enabled viewing of convolution weights. Still needs some ironing out of the details of the model. Test and implement multiple hidden units of the same type. The test of the convolutional rbm is going well on the lattice proteins. You can tell it is working to certain degree

+++ rbm_torch/rbm_utils.py
@@ -74,8 +74,8 @@ aadict['.'] = -1
 
 def get_ax(ax, i, nrows, ncols):
     if (ncols > 1) & (nrows > 1):
-        col = i % ncols
-        row = i / ncols
+        col = int(i % ncols)
+        row = int(i / ncols)
         ax_ = ax[row, col]
     elif (ncols > 1) & (nrows == 1):
         ax_ = ax[i]
